Try UTF-8 when converting an Exception to unicode
The motivation is that there is no way to guess the encoding in the
exception's message but our own messages (generate by our C++ libs) tend
to be in UTF-8. So first try to decode as UTF-8, which is strict and will
complain the loudest, and then fallback to the locale's preferred
encoding.

This should keep 99% of the previous usecases unaffected but give nice
tracebacks/error reports for exceptions coming from C++ with non-ascii in
their messages.

ROCKY-2099